+++ src/utils/debug.ts
@@ -1,5 +1,0 @@
-import * as debug from 'debug';
-
-const getDebug = (identifier: string) => debug(`oxid:${identifier}`);
-
-export { getDebug };